Added database connection and improved formatting

+++ src/pages/courses/courses_BL.php
@@ -7,22 +7,28 @@
  * Constants:
  */
 
-if(isset($_POST['submit'])){
+if (!file_exists(DATABASE_MODULE) || !is_readable(DATABASE_MODULE)) {
+    exit(DATABASE_MODULE_NOT_FOUND);
+}
+
+$DB = connect();
+
+if (isset($_POST['submit'])) {
     $query = "INSERT INTO courses (courseName) VALUES ('".$_POST['newCourse']."')";
     mysqli_query($DB, $query);
 }
 
-if(isset($_POST['delete'])){
+if (isset($_POST['delete'])) {
     $query = "DELETE FROM teachersToSubject WHERE subjectId=".$_POST['delete'][0].";";
     mysqli_query($DB, $query);
     $query = "DELETE FROM subjects WHERE subjectId=".$_POST['delete'][0].";";
     mysqli_query($DB, $query);
 }
 
-if(isset($_POST['deleteCourse'])){
-    foreach($_POST['deleteCourse'] as $deleteCourse){
+if (isset($_POST['deleteCourse'])) {
+    foreach ($_POST['deleteCourse'] as $deleteCourse) {
         $subjects = mysqli_query($DB, "SELECT subjectId FROM subjects WHERE courseId=".$deleteCourse.';');
-        while($subject = mysqli_fetch_array($subjects)){
+        while ($subject = mysqli_fetch_array($subjects)) {
             $query = "DELETE FROM teachersToSubject WHERE=".$subject['subjectId'].";";
             mysqli_query($DB, $query);
             $query = "DELETE FROM subjects WHERE subjectId=".$subject['subjectId'].";";
